% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmaskr_data.R
\docType{data}
\name{countmaskr_data}
\alias{countmaskr_data}
\title{countmaskr dataset}
\format{
\subsection{\code{countmaskr_data}}{

A data frame with 1500 rows and 6 columns:
\describe{
\item{id}{Subject id}
\item{age}{Age in years as integer}
\item{gender}{Administrative sex}
\item{race}{Race}
\item{ethnicity}{Ethnicity}
\item{age_group}{Age as categorical variable}
...
}
}
}
\usage{
data(countmaskr_data)
}
\description{
A synthetic dataset of 1500 subjects with demographic information ascertaining atleast one category from the demographics has counts less than 11.
Report ...
}
\keyword{datasets}
