% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_masked_counts.R
\name{get_masked_counts}
\alias{get_masked_counts}
\title{Function to perform threshold based cell masking}
\usage{
get_masked_counts(x, threshold = 11)
}
\arguments{
\item{x}{vector of length N}

\item{threshold}{threshold below with the values must be suppressed}
}
\value{
a character vector with primary and/or secondary masked cell
}
\description{
Function to perform threshold based cell masking
}
\examples{
x = c(5, 11, 43, 55, 65, 121, 1213,0,NA)
threshold_suppressor(x)
}
