% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_percentages.R
\name{get_masked_percentages}
\alias{get_masked_percentages}
\title{Function to generate masked percentage for given numeric vector}
\usage{
get_masked_percentages(x, x_masked = NULL, threshold = 11)
}
\arguments{
\item{x}{vector of length N}

\item{x_masked}{masked vector of parameter x. Defaulted to NULL}

\item{threshold}{threshold of defined count suppression}
}
\value{
character vector of masked percentages
}
\description{
Function to generate masked percentage for given numeric vector
}
