% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_masked_table.R
\name{get_masked_table}
\alias{get_masked_table}
\title{Function to apply cell suppression methods on a table}
\usage{
get_masked_table(
  data,
  threshold = 11,
  col_groups,
  group_by = NULL,
  overwrite_columns = T,
  percentages = F
)
}
\arguments{
\item{data}{input data with a column that has counts which need suppression}

\item{threshold}{threshold value for suppression for the threshold_suppression() function. defaulted to 11}

\item{col_groups}{columns that requires suppression. If two way tables. enter columns groups that require row-wise suppression as list()}

\item{group_by}{variable name to group the masking by.}
}
\value{
a tibble with row and column wise masking. masked columns will return as character vector
}
\description{
Function to apply cell suppression methods on a table
}
\examples{
df <- tibble::tribble(
  ~block, ~Characteristics, ~col1, ~col2, ~col3, ~col4,
  "total", "Patient Totals", 222, 1434, 525, 8,
  "sex", "Male", 190, 1407, 8, 2,
  "sex", "Female", 17, 20, 511, 2,
  "sex", "Sex - Other", 15, 7, 6, 4,
  "race", "White", 102, 1385, 75, 1,
  "race", "African American / Black", 75, 30, 325, 0,
  "race", "Asian", 20, 9, 100, 2,
  "race", "Native American / Pacific Islander", 15, 10, 4, 3,
  "race", "Race - Other", 10, 0, 21, 2,
  "Presence of Diabetes", "Presence of Diabetes", 215, 6, 215, 0,
) \%>\%
  mutate(
    aggr_all_cols = col1 + col2 + col3 + col4,
    aggr_col1_col2 = col1 + col2,
    aggr_col3_col4 = col3 + col4
  )
mask_table(df,
  group_by = "block",
  col_groups = list(
    c("aggr_col1_col2", "col1", "col2"),
    c("aggr_col3_col4", "col3", "col4")
  )
)

}
