% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_gtsummary.R
\name{mask_gtsummary}
\alias{mask_gtsummary}
\title{Function to mask output from \code{gtsummary::tbl_summary()} function}
\usage{
mask_gtsummary(gttbl)
}
\arguments{
\item{gttbl}{output from gtsummary::tbl_summary()}
}
\value{
A masked tbl_summary object
}
\description{
This is a wrapper around the gtsummary::tbl_summary()`= function to produce a masked version of the aggregated table.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
data(countmaskr_data)
mask_gtsummary(gtsummary::tbl_summary(countmaskr_data[,-1]))
}
