% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_percentages.R
\name{mask_percentages}
\alias{mask_percentages}
\title{Function to generate masked percentage for given numeric vector}
\usage{
mask_percentages(x, x_masked = NULL, threshold = 11)
}
\arguments{
\item{x}{vector of length N}

\item{x_masked}{masked vector of parameter x. Defaulted to NULL}

\item{threshold}{threshold of defined count suppression}
}
\value{
character vector of masked percentages
}
\description{
This function converts a vector to percentages
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
x <- c(5, 11, 43, 55, 65, 121, 1213, 0, NA)
mask_percentages(x)
}
