% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_small_cells.R
\name{mask_small_cells}
\alias{mask_small_cells}
\title{Small Cell Suppression function}
\usage{
mask_small_cells(data, threshold = 11, cols)
}
\arguments{
\item{data}{Takes in a dataset with a required column named `block`.}

\item{threshold}{Value below which the cells with be suppressed}

\item{cols}{col names that require suppression}
}
\value{
dataset with suppressed values for the columns specified in @param cols
}
\description{
Small Cell Suppression function
}
\examples{

df_row_masked <- mask_small_cells(df, cols = c('col1','col2'))

}
