% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_small_cell_2.R
\name{mask_small_cells_2}
\alias{mask_small_cells_2}
\title{Function to mask small. cell in a vector}
\usage{
mask_small_cells_2(x, threshold = 11)
}
\arguments{
\item{x}{numeric vector of length N}

\item{threshold}{threshold for small cell aka 'primary cell'}
}
\value{
vector with masked values for primary small cell and secondary non-small cell where required
}
\description{
Function to mask small. cell in a vector
}
\examples{
mask_small_cells(x = c(102,74,30,30,4,NA))

df <- tibble::tribble(
~block,                      ~Characterstics, ~col1, ~col2,
"sex",                               "Male",   190,  1407,
"sex",                             "Female",    17,    20,
"sex",                        "Sex - Other",    15,     7,
"race",                              "White",   102,  1385,
"race",           "African American / Black",    75,    20,
"race",                              "Asian",    20,    19,
"race", "Native American / Pacific Islander",    15,    10,
"race",                       "Race - Other",    10,     0,
"total",                     "Patient Totals",   222,  1434,
"Presence of Diabetes","Presence of Diabetes",   215,     6
)
df \%>\% group_by(block) \%>\% mutate(across(starts_with('col'), ~mask_small_cells(.))

}
