% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_suppression.R
\name{rowwise_suppression}
\alias{rowwise_suppression}
\title{Column-wise suppression}
\usage{
rowwise_suppression(data, cols)
}
\arguments{
\item{data}{Output dataset from (suppress_small_cells())}

\item{cols}{groups of columns that needs a column-wise suppression. This parameter should be passed as a list object. If multiple groups exists,
each group can be a seperate list object.}
}
\value{
dataset with column-wise suppressed cells
}
\description{
Column-wise suppression
}
\examples{

df_row_col_masked <- colwise_suppression(df_row_masked, cols = list(c('col1_masked','col2_masked')))
}
