% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmaskr-package.R
\docType{package}
\name{countmaskr-package}
\alias{countmaskr}
\alias{countmaskr-package}
\title{countmaskr: Small Cell Masking Tool for One- & Two-Way Tabular Reports}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides automated small-cell suppression for one- and two-way frequency tables. Cells falling below a user-defined frequency threshold are masked, with suppression propagated to secondary cells to prevent indirect disclosure. Designed for clinical and health administrative data, the package supports a range of tabular structures and fits into reproducible reporting pipelines, reducing manual review while applying consistent suppression rules across data sharing workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://query-fulfillment.github.io/countmaskr/}
}

}
\author{
\strong{Maintainer}: Sahal Master \email{sahalmaster@outlook.com} (\href{https://orcid.org/0000-0002-8518-8107}{ORCID})

Authors:
\itemize{
  \item Amy Goodwin Davies \email{goodwindaa@chop.edu}
  \item Allison Zelinski \email{zelinskia@chop.edu}
  \item Qiwei Shen \email{shenq@chop.edu} (\href{https://orcid.org/0009-0003-0631-0026}{ORCID})
}

Other contributors:
\itemize{
  \item Nicole Marchesani \email{MARCHESANN@chop.edu} (\href{https://orcid.org/0009-0006-0683-7575}{ORCID}) [contributor]
  \item Aqsa Khan \email{khana8@chop.edu} (\href{https://orcid.org/0009-0000-9574-4704}{ORCID}) [contributor]
  \item Rhonda DeCook \email{rhonda-decook@uiowa.edu} [contributor]
}

}
\keyword{internal}
