% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmaskr-package.R
\docType{package}
\name{countmaskr-package}
\alias{countmaskr}
\alias{countmaskr-package}
\title{countmaskr: Small Cell Masking Tool for One- & Two-Way Tabular Reports}
\description{
This is an R package designed to address patient privacy concerns by automating the masking of cell counts in aggregated tabular reports. This user-friendly and versatile tool applies threshold-based masking to primary, secondary, and tertiary cells, replacing true counts with masked values. It can handle various report types, saving significant time and reducing privacy risks. This package ensures the protection of patient data while retaining accurate counts in complex tabular reports, making it a valuable addition to data sharing and privacy protection workflows.
}
\author{
\strong{Maintainer}: Sahal Master \email{sahalmaster@outlook.com} (\href{https://orcid.org/0000-0002-8518-8107}{ORCID})

Authors:
\itemize{
  \item Amy Goodwin Davies \email{goodwindaa@chop.edu}
  \item Ally Zelinski \email{zelinskia@chop.edu}
  \item Qiwei Shen \email{shenq@chop.edu}
}

}
\keyword{internal}
