% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_counts.R
\name{mask_counts}
\alias{mask_counts}
\title{Function to perform threshold based cell masking - method 1}
\usage{
mask_counts(x, threshold = 11)
}
\arguments{
\item{x}{vector of length N}

\item{threshold}{threshold below with the values must be suppressed}
}
\value{
a character vector with primary and/or secondary masked cell
}
\description{
This function is a workhorse of the mask_table function. It identifies primary and secondary cells, and masks them as necessary
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
x1 <- c(5, 11, 43, 55, 65, 121, 1213, 0, NA)
x2 <- c(1,1,1,55, 65, 121, 1213, 0, NA)
x3 <- c(11, 10,10, 55, 65, 121, 1213, 0, NA)

mask_counts(x)

 df <- tibble::tribble(
~block, ~Characteristics, ~group1, ~group2, ~group3, ~group4,
"sex", "Male", 190, 1407, 8, 2,
"sex", "Female", 17, 20, 511, 2,
"sex", "Sex - Other", 15, 7, 6, 4,
"race", "White", 102, 1385, 75, 1,
"race", "African American / Black", 75, 30, 325, 0,
"race", "Asian", 20, 9, 100, 2,
"race", "Native American / Pacific Islander", 15, 10, 4, 3,
"race", "Race - Other", 10, 0, 21, 2,
) \%>\%
  dplyr::mutate(
    aggr_group_all = group1 + group2 + group3 + group4,
    aggr_group_1_2 = group1 + group2,
    aggr_group_3_4 = group3 + group4
 )

 df \%>\% group_by(block) \%>\%
    mutate(across(contains('group'), ~mask_counts(.),.names = "{col}_masked"))
}
