% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_table.R
\name{mask_table}
\alias{mask_table}
\title{Function to apply cell suppression methods on a two-by-two tables}
\usage{
mask_table(
  data,
  threshold = 11,
  col_groups,
  group_by = NULL,
  overwrite_columns = T,
  percentages = F
)
}
\arguments{
\item{data}{input data with a column that has counts which need suppression}

\item{threshold}{threshold value for suppression for the threshold_suppression() function. defaulted to 11}

\item{col_groups}{columns that requires suppression. If two way tables. enter columns groups that require row-wise suppression as list()}

\item{group_by}{variable name to group the masking by}

\item{overwrite_columns}{Boolean parameter to overwrite columns}

\item{percentages}{Boolean parameter for generate masked percentages. naming convention will be as follows: ('{col}_perc_masked')}
}
\value{
a tibble with row and column wise masking. masked columns will return as character vector
}
\description{
This is a multi-tasking function to convert a simple table 1 with counts and percentages both  -masked and unmasked
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
df <- tibble::tribble(
  ~block, ~Characteristics, ~group1, ~group2, ~group3, ~group4,
  "sex", "Male", 190, 1407, 8, 2,
  "sex", "Female", 17, 20, 511, 2,
  "sex", "Sex - Other", 15, 7, 6, 4,
  "race", "White", 102, 1385, 75, 1,
  "race", "African American / Black", 75, 30, 325, 0,
  "race", "Asian", 20, 9, 100, 2,
  "race", "Native American / Pacific Islander", 15, 10, 4, 3,
  "race", "Race - Other", 10, 0, 21, 2,
) \%>\%
  dplyr::mutate(
    aggr_group_all = group1 + group2 + group3 + group4,
    aggr_group_1_2 = group1 + group2,
    aggr_group_3_4 = group3 + group4
  )
mask_table(df,
  group_by = "block",
  col_groups = list(
    c("aggr_group_1_2", "group1", "group2"),
    c("aggr_group_3_4", "group3", "group4")
  ),
  overwrite_columns = FALSE,
  percentages = TRUE
)
}
