% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb_counts.R
\name{perturb_counts}
\alias{perturb_counts}
\title{Function to perturb counts in a vector with small cells}
\usage{
perturb_counts(x, threshold = 10)
}
\arguments{
\item{x}{numeric vector of length N}

\item{threshold}{threshold for small cell aka 'primary cell'}
}
\value{
a vector with added weighted noise to the non-primary cells
If the non-secondary cell do not have enough room to distribute noise, the function will return with an error
suggesting to use Threshold-based suppression.
}
\description{
This is a function to perturb counts in a vector where only one primary cell is present and requires
secondary cell to be masked - Algorithm 3 (A3)
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
x1 <- c(5, 11, 43, 55, 65, 121, 1213, 0, NA)
x2 <- c(1, 1, 1, 55, 65, 121, 1213, 0, NA)
x3 <- c(11, 10, 10, 55, 65, 121, 1213, 0, NA)

lapply(list(x1, x2, x3), perturb_counts)

data('countmaskr_data')

aggregate_table <- countmaskr_data \%>\%
  select(-c(id, age)) \%>\%
  gather(block, Characteristics) \%>\%
  group_by(block, Characteristics) \%>\%
  summarise(N = n()) \%>\%
  ungroup()

aggregate_table \%>\%
  group_by(block) \%>\%
  mutate(N_masked = perturb_counts(N))

}
